Move the SpeedDial to its own component
